pyproject.toml
[tool.poetry]
name = "soa-bridge-match"
version = "0.1.0"
description = "Dataset builder"
authors = ["Geoff Low <glow@mdsol.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"fhir.resources" = "^6.2.2"
xport = "^3.6.1"
jupyterlab = "^3.3.4"
openpyxl = "^3.0.9"
python-dotenv = "^0.20.0"
Faker = "^13.12.0"
pytz = "^2022.2.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
